robotops-grafana/gateway/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
paho-mqtt>=2.0.0
prometheus-client>=0.19.0
requests>=2.31.0
python-dotenv>=1.0.0
